refactor: prepare miden-mast-package for use as a dependency of miden-assembly
