+++ assembly/Cargo.toml
@@ -30,6 +30,7 @@ miden-core = { path = "../core", version = "0.15", default-features = false, fea
     "diagnostics",
 ] }
 miden-assembly-syntax = { path = "../syntax", version = "0.15", default-features = false }
+miden-mast-package = { path = "../package", version = "0.15", default-features = false }
 smallvec = { version = "1.13", features = [
     "union",
     "const_generics",
@@ -38,4 +39,11 @@ smallvec = { version = "1.13", features = [
 thiserror.workspace = true
 
 [dev-dependencies]
+miden-mast-package = { path = "../package", version = "0.15", features = [
+    "arbitrary",
+] }
 pretty_assertions = "1.4"
+proptest = { version = "1.6", default-features = false, features = [
+    "no_std",
+    "alloc",
+] }
