+++ package/Cargo.toml
@@ -17,12 +17,17 @@ edition.workspace = true
 bench = false
 doctest = false
 
+[features]
+default = []
+arbitrary = [
+    "dep:proptest-derive",
+    "dep:proptest",
+    "miden-assembly-syntax/arbitrary",
+]
+
 [dependencies]
 derive_more = { version = "2.0.1", features = ["from"] }
 miden-assembly-syntax = { path = "../syntax", version = "0.15", default-features = false }
 miden-core = { path = "../core", version = "0.15", default-features = false }
-
-[dev-dependencies]
-miden-assembly = { path = "../assembly", version = "0.15", default-features = false }
-proptest = "1.6"
-proptest-derive = "0.5"
+proptest = { version = "1.6", optional = true }
+proptest-derive = { version = "0.5", optional = true }
