+++ package/src/dependency/mod.rs
@@ -10,7 +10,7 @@ pub(crate) mod resolver;
 
 /// The name of a dependency
 #[derive(Debug, Clone, PartialEq, Eq, derive_more::From)]
-#[cfg_attr(test, derive(proptest_derive::Arbitrary))]
+#[cfg_attr(feature = "arbitrary", derive(proptest_derive::Arbitrary))]
 pub struct DependencyName(String);
 
 impl Serializable for DependencyName {
@@ -28,14 +28,14 @@ impl Deserializable for DependencyName {
 
 /// A package dependency
 #[derive(Debug, Clone, PartialEq, Eq)]
-#[cfg_attr(test, derive(proptest_derive::Arbitrary))]
+#[cfg_attr(feature = "arbitrary", derive(proptest_derive::Arbitrary))]
 pub struct Dependency {
     /// The name of the dependency.
     /// Serves as a human-readable identifier for the dependency and a search hint for the resolver
     pub name: DependencyName,
     /// The digest of the dependency.
     /// Serves as an ultimate source of truth for identifying the dependency.
-    #[cfg_attr(test, proptest(value = "Word::default()"))]
+    #[cfg_attr(feature = "arbitrary", proptest(value = "Word::default()"))]
     pub digest: Word,
 }
